requirements.txt
fastapi==0.110.1
pymongo==4.6.3
python-dotenv==1.0.1
uvicorn==0.29.0
